[RDPF-187] 리뷰 페이지 조회 타입 수정

+++ perfume-api/src/main/java/io/perfume/api/review/application/facade/ReviewDetailFacadeService.java
@@ -37,7 +37,7 @@ public class ReviewDetailFacadeService {
     this.getReviewTagUseCase = getReviewTagUseCase;
   }
 
-  public List<ReviewDetailResult> getPaginatedReviews(int page, int size) {
+  public List<ReviewDetailResult> getPaginatedReviews(long page, long size) {
     final var reviews =
         getReviewsUseCase
             .getPaginatedReviews(page, size)
