+++ perfume-api/src/main/java/io/perfume/api/review/application/service/GetReviewsService.java
@@ -16,7 +16,7 @@ public class GetReviewsService implements GetReviewsUseCase {
   }
 
   @Override
-  public List<ReviewResult> getPaginatedReviews(int page, int size) {
+  public List<ReviewResult> getPaginatedReviews(long page, long size) {
     return reviewQueryRepository
         .findByPage(page, size)
         .stream()
